// apps/web/generated/api/models/formFieldResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Freelance Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { FormFieldOptionResponse } from './formFieldOptionResponse';
import type { FormFieldResponseValidationRules } from './formFieldResponseValidationRules';
import type { FormFieldType } from './formFieldType';

export interface FormFieldResponse {
  field_id: string;
  field_key: string;
  label: string;
  description: string | null;
  field_type: FormFieldType;
  is_required: boolean;
  is_repeatable: boolean;
  is_unique: boolean;
  sort_order: number;
  validation_rules: FormFieldResponseValidationRules;
  is_active: boolean;
  options: FormFieldOptionResponse[];
}