// apps/web/generated/api/models/projectResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Freelance Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { BudgetResponse } from './budgetResponse';
import type { FormValueInputRequest } from './formValueInputRequest';
import type { FreelancerLevelEnum } from './freelancerLevelEnum';
import type { ProjectPriority } from './projectPriority';
import type { ProjectStatus } from './projectStatus';
import type { ProjectVisibility } from './projectVisibility';

export interface ProjectResponse {
  project_id: string;
  project_code: string;
  customer_user_id: string;
  category_id: string;
  form_template_id: string;
  form_values?: FormValueInputRequest[];
  required_level: FreelancerLevelEnum | null;
  title: string;
  description: string;
  status: ProjectStatus;
  visibility: ProjectVisibility;
  priority: ProjectPriority;
  budget: BudgetResponse;
  assigned_supervisor_user_id: string | null;
  selected_application_id: string | null;
  application_deadline: string | null;
  created_by_user_id: string | null;
  created_at: string;
}